:Bug: Long to Int

+++ src/main/kotlin/com/toyProject7/karrot/article/persistence/ArticleRepository.kt
@@ -20,7 +20,7 @@ interface ArticleRepository : JpaRepository<ArticleEntity, Long> {
 
     @Modifying
     @Query("UPDATE articles a SET a.viewCount = a.viewCount + 1 WHERE a.id = :id")
-    fun incrementViewCount(id: Long): Long
+    fun incrementViewCount(id: Long): Int
 
     fun countBySellerId(id: String): Int
 }
